requirements.txt
asttokens==3.0.0
colorama==0.4.6
comm==0.2.2
debugpy==1.8.14
decorator==5.2.1
exceptiongroup==1.3.0
executing==2.2.0
importlib-metadata==8.7.0
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.2
jupyter-client==8.6.3
jupyter-core==5.7.2
matplotlib-inline==0.1.7
mutagen==1.47.0
nest-asyncio==1.6.0
numpy==2.0.2
packaging==25.0
parso==0.8.4
pillow==11.2.1
platformdirs==4.3.8
playsound==1.3.0
prompt-toolkit==3.0.51
psutil==7.0.0
pure-eval==0.2.3
pygame==2.6.1
pygments==2.19.1
python-dateutil==2.9.0.post0
pywin32==310
pyzmq==26.4.0
six==1.17.0
stack-data==0.6.3
tkinter-tooltip==3.1.2
tornado==6.5
traitlets==5.14.3
typing-extensions==4.13.2
wcwidth==0.2.13
zipp==3.21.0